+++ app/Http/Controllers/OrderController.php
@@ -8,6 +8,7 @@ use App\Models\Cart;
 use App\Models\Order;
 use App\Models\Shipping;
 use App\User;
+use Illuminate\Support\Facades\Gate;
 use PDF;
 use Notification;
 use Helper;
@@ -23,6 +24,10 @@ class OrderController extends Controller
      */
     public function index()
     {
+        if(Gate::denies('list_order'))
+        {
+            abort('403');
+        }
         $orders = Order::orderBy('created_at', 'DESC')->paginate(10);
         return view('backend.order.index')->with('orders', $orders);
     }
@@ -162,6 +167,10 @@ class OrderController extends Controller
      */
     public function edit($id)
     {
+        if(Gate::denies('update_order'))
+        {
+            abort('403');
+        }
         $order = Order::find($id);
         return view('backend.order.edit')->with('order', $order);
     }
@@ -209,6 +218,10 @@ class OrderController extends Controller
      */
     public function destroy($id)
     {
+        if(Gate::denies('delete_order'))
+        {
+            abort('403');
+        }
         $order = Order::find($id);
         if ($order) {
             $status = $order->delete();
@@ -267,7 +280,6 @@ class OrderController extends Controller
     public function incomeChart(Request $request)
     {
         $year = \Carbon\Carbon::now()->year;
-        // dd($year);
         $items = Order::with(['cart_info'])->whereYear('created_at', $year)->where('status', 'delivered')->get()
             ->groupBy(function ($d) {
                 return \Carbon\Carbon::parse($d->created_at)->format('m');
@@ -276,9 +288,7 @@ class OrderController extends Controller
         foreach ($items as $month => $item_collections) {
             foreach ($item_collections as $item) {
                 $amount = $item->cart_info->sum('amount');
-                // dd($amount);
                 $m = intval($month);
-                // return $m;
                 isset($result[$m]) ? $result[$m] += $amount : $result[$m] = $amount;
             }
         }
